fix ux + add interface for google storage / OCI
