+++ web/src/components/icons/icons.tsx
@@ -43,8 +43,9 @@ import {
 import { SiBookstack } from "react-icons/si";
 import Image from "next/image";
 import jiraSVG from "../../../public/Jira.svg";
-import blobSVG from "../../../public/blob.svg";
 import confluenceSVG from "../../../public/Confluence.svg";
+import OCIStorageSVG from "../../../public/OCI.svg";
+import googleCloudStorageIcon from "../../../public/GoogleCloudStorage.png";
 import guruIcon from "../../../public/Guru.svg";
 import gongIcon from "../../../public/Gong.png";
 import requestTrackerIcon from "../../../public/RequestTracker.png";
@@ -426,6 +427,20 @@ export const ConfluenceIcon = ({
   );
 };
 
+export const OCIStorageIcon = ({
+  size = 16,
+  className = defaultTailwindCSS,
+}: IconProps) => {
+  return (
+    <div
+      style={{ width: `${size + 4}px`, height: `${size + 4}px` }}
+      className={`w-[${size + 4}px] h-[${size + 4}px] -m-0.5 ` + className}
+    >
+      <Image src={OCIStorageSVG} alt="Logo" width="96" height="96" />
+    </div>
+  );
+};
+
 export const JiraIcon = ({
   size = 16,
   className = defaultTailwindCSS,
@@ -441,21 +456,6 @@ export const JiraIcon = ({
   );
 };
 
-export const BlobIcon = ({
-  size = 16,
-  className = defaultTailwindCSS,
-}: IconProps) => {
-  // Jira Icon has a bit more surrounding whitespace than other icons, which is why we need to adjust it here
-  return (
-    <div
-      style={{ width: `${size + 4}px`, height: `${size + 4}px` }}
-      className={`w-[${size + 4}px] h-[${size + 4}px] -m-0.5 ` + className}
-    >
-      <Image src={blobSVG} alt="Logo" width="96" height="96" />
-    </div>
-  );
-};
-
 export const ZulipIcon = ({
   size = 16,
   className = defaultTailwindCSS,
@@ -466,6 +466,20 @@ export const ZulipIcon = ({
       className={`w-[${size}px] h-[${size}px] ` + className}
     >
       <Image src={zulipIcon} alt="Logo" width="96" height="96" />
+    </div>
+  );
+};
+
+export const GoogleStorageIcon = ({
+  size = 16,
+  className = defaultTailwindCSS,
+}: IconProps) => {
+  return (
+    <div
+      style={{ width: `${size + 4}px`, height: `${size + 4}px` }}
+      className={`w-[${size + 4}px] h-[${size + 4}px] -m-0.5 ` + className}
+    >
+      <Image src={googleCloudStorageIcon} alt="Logo" width="96" height="96" />
     </div>
   );
 };
